+++ Roles/Impostor/Penguin.cs
@@ -43,12 +43,8 @@ namespace EHR.Impostor
         public static bool IsVictim(PlayerControl pc)
         {
             foreach (KeyValuePair<byte, PlayerState> state in Main.PlayerStates)
-            {
                 if (state.Value.Role is Penguin { IsEnable: true, VictimCanUseAbilities: false } pg && pg.AbductVictim != null && pg.AbductVictim.PlayerId == pc.PlayerId)
-                {
                     return true;
-                }
-            }
 
             return false;
         }
@@ -56,14 +52,18 @@ namespace EHR.Impostor
         public override void SetupCustomOption()
         {
             Options.SetupRoleOptions(Id, TabGroup.ImpostorRoles, CustomRoles.Penguin);
+
             OptionAbductTimerLimit = new FloatOptionItem(Id + 11, "PenguinAbductTimerLimit", new(1f, 20f, 1f), 10f, TabGroup.ImpostorRoles)
                 .SetParent(Options.CustomRoleSpawnChances[CustomRoles.Penguin])
                 .SetValueFormat(OptionFormat.Seconds);
+
             OptionMeetingKill = new BooleanOptionItem(Id + 12, "PenguinMeetingKill", false, TabGroup.ImpostorRoles)
                 .SetParent(Options.CustomRoleSpawnChances[CustomRoles.Penguin]);
+
             OptionSpeedDuringDrag = new FloatOptionItem(Id + 13, "PenguinSpeedDuringDrag", new(0.1f, 3f, 0.1f), 1f, TabGroup.ImpostorRoles)
                 .SetParent(Options.CustomRoleSpawnChances[CustomRoles.Penguin])
                 .SetValueFormat(OptionFormat.Multiplier);
+
             OptionVictimCanUseAbilities = new BooleanOptionItem(Id + 14, "PenguinVictimCanUseAbilities", false, TabGroup.ImpostorRoles)
                 .SetParent(Options.CustomRoleSpawnChances[CustomRoles.Penguin]);
         }
@@ -127,18 +127,12 @@ namespace EHR.Impostor
 
         public override void ApplyGameOptions(IGameOptions opt, byte playerId)
         {
-            if (IsGoose)
-            {
-                opt.SetVision(false);
-            }
+            if (IsGoose) opt.SetVision(false);
         }
 
         private void SendRPC()
         {
-            if (!IsEnable || !Utils.DoRPC)
-            {
-                return;
-            }
+            if (!IsEnable || !Utils.DoRPC) return;
 
             MessageWriter writer = AmongUsClient.Instance.StartRpcImmediately(PlayerControl.LocalPlayer.NetId, (byte)CustomRPC.PenguinSync, SendOption.Reliable);
             writer.Write(PenguinId);
@@ -168,10 +162,7 @@ namespace EHR.Impostor
 
         private void AddVictim(PlayerControl target)
         {
-            if (!IsEnable)
-            {
-                return;
-            }
+            if (!IsEnable) return;
 
             AbductVictim = target;
             AbductTimer = AbductTimerLimit;
@@ -184,10 +175,7 @@ namespace EHR.Impostor
 
         private void RemoveVictim(bool allowDelay = true)
         {
-            if (!IsEnable)
-            {
-                return;
-            }
+            if (!IsEnable) return;
 
             if (!allowDelay)
             {
@@ -217,12 +205,10 @@ namespace EHR.Impostor
 
         public override bool OnCheckMurder(PlayerControl killer, PlayerControl target)
         {
-            if (!IsEnable)
-            {
-                return true;
-            }
+            if (!IsEnable) return true;
 
             bool doKill = !IsGoose;
+
             if (AbductVictim != null)
             {
                 if (target.PlayerId != AbductVictim.PlayerId)
@@ -241,19 +227,13 @@ namespace EHR.Impostor
             }
             else
             {
-                if (!IsGoose && !killer.RpcCheckAndMurder(target, true))
-                {
-                    return false;
-                }
+                if (!IsGoose && !killer.RpcCheckAndMurder(target, true)) return false;
 
                 doKill = false;
                 AddVictim(target);
             }
 
-            if (doKill)
-            {
-                killer.Kill(target);
-            }
+            if (doKill) killer.Kill(target);
 
             return false;
         }
@@ -265,39 +245,25 @@ namespace EHR.Impostor
 
         public override void OnReportDeadBody()
         {
-            if (!IsEnable)
-            {
-                return;
-            }
+            if (!IsEnable) return;
 
             stopCount = true;
+
             // If you meet a meeting with time running out, kill it even if you're on a ladder.
             if (AbductVictim != null && AbductTimer <= 0f)
             {
-                if (!IsGoose)
-                {
-                    Penguin_.Kill(AbductVictim);
-                }
+                if (!IsGoose) Penguin_.Kill(AbductVictim);
 
                 RemoveVictim(false);
             }
 
             if (MeetingKill)
             {
-                if (!AmongUsClient.Instance.AmHost)
-                {
-                    return;
-                }
+                if (!AmongUsClient.Instance.AmHost) return;
 
-                if (AbductVictim == null)
-                {
-                    return;
-                }
+                if (AbductVictim == null) return;
 
-                if (!IsGoose)
-                {
-                    Penguin_.Kill(AbductVictim);
-                }
+                if (!IsGoose) Penguin_.Kill(AbductVictim);
 
                 RemoveVictim();
             }
@@ -305,10 +271,7 @@ namespace EHR.Impostor
 
         public override void AfterMeetingTasks()
         {
-            if (Main.NormalOptions.MapId == 4)
-            {
-                return;
-            }
+            if (Main.NormalOptions.MapId == 4) return;
 
             //Maps other than Airship
             RestartAbduct();
@@ -317,51 +280,33 @@ namespace EHR.Impostor
         public static void OnSpawnAirship()
         {
             foreach (KeyValuePair<byte, PlayerState> state in Main.PlayerStates)
-            {
                 if (state.Value.Role is Penguin { IsEnable: true } pg)
-                {
                     pg.RestartAbduct();
-                }
-            }
         }
 
         private void RestartAbduct()
         {
-            if (!IsEnable)
-            {
-                return;
-            }
+            if (!IsEnable) return;
 
-            if (AbductVictim != null)
-            {
-                Penguin_.MarkDirtySettings();
-            }
+            if (AbductVictim != null) Penguin_.MarkDirtySettings();
 
             stopCount = false;
         }
 
         public override void OnFixedUpdate(PlayerControl pc)
         {
-            if (!IsEnable)
-            {
-                return;
-            }
+            if (!IsEnable) return;
 
-            if (!AmongUsClient.Instance.AmHost)
-            {
-                return;
-            }
+            if (!AmongUsClient.Instance.AmHost) return;
 
-            if (!GameStates.IsInTask)
-            {
-                return;
-            }
+            if (!GameStates.IsInTask) return;
 
             if (!stopCount)
             {
                 AbductTimer -= Time.fixedDeltaTime;
 
                 Count++;
+
                 if (Count >= 10)
                 {
                     Count = 0;
@@ -396,31 +341,35 @@ namespace EHR.Impostor
                     if (!AbductVictim.MyPhysics.Animations.IsPlayingAnyLadderAnimation())
                     {
                         PlayerControl abductVictim = AbductVictim;
+
                         LateTask.New(() =>
                         {
-                            if (IsGoose)
-                            {
-                                return;
-                            }
+                            if (IsGoose) return;
 
                             int sId = abductVictim.NetTransform.lastSequenceId + 5;
                             abductVictim.NetTransform.SnapTo(Penguin_.transform.position, (ushort)sId);
                             Penguin_.Kill(abductVictim);
 
-                            CustomRpcSender sender = CustomRpcSender.Create("PenguinMurder");
+                            var sender = CustomRpcSender.Create("PenguinMurder");
+
                             {
                                 sender.AutoStartRpc(abductVictim.NetTransform.NetId, (byte)RpcCalls.SnapTo);
+
                                 {
                                     NetHelpers.WriteVector2(Penguin_.transform.position, sender.stream);
                                     sender.Write(abductVictim.NetTransform.lastSequenceId);
                                 }
+
                                 sender.EndRpc();
                                 sender.AutoStartRpc(Penguin_.NetId, (byte)RpcCalls.MurderPlayer);
+
                                 {
                                     sender.WriteNetObject(abductVictim);
                                 }
+
                                 sender.EndRpc();
                             }
+
                             sender.SendMessage();
                         }, 0.3f, "PenguinMurder");
 
@@ -431,39 +380,26 @@ namespace EHR.Impostor
                 else if (!AbductVictim.MyPhysics.Animations.IsPlayingAnyLadderAnimation())
                 {
                     Vector3 position = Penguin_.transform.position;
+
                     if (!Penguin_.IsHost())
-                    {
                         Utils.TP(AbductVictim.NetTransform, position, log: false);
-                    }
                     else
                     {
                         LateTask.New(() =>
                         {
-                            if (AbductVictim != null)
-                            {
-                                Utils.TP(AbductVictim.NetTransform, position, log: false);
-                            }
+                            if (AbductVictim != null) Utils.TP(AbductVictim.NetTransform, position, log: false);
                         }, 0.25f, log: false);
                     }
                 }
             }
-            else if (AbductTimer <= 100f)
-            {
-                AbductTimer = 255f;
-            }
+            else if (AbductTimer <= 100f) AbductTimer = 255f;
         }
 
         public override string GetSuffix(PlayerControl seer, PlayerControl target, bool hud = false, bool meeting = false)
         {
-            if (seer == null || seer.PlayerId != target.PlayerId || (seer.IsModClient() && !hud))
-            {
-                return string.Empty;
-            }
+            if (seer == null || seer.PlayerId != target.PlayerId || (seer.IsModClient() && !hud)) return string.Empty;
 
-            if (!Main.PlayerStates.TryGetValue(seer.PlayerId, out PlayerState state) || state.Role is not Penguin pg || pg.AbductVictim == null)
-            {
-                return string.Empty;
-            }
+            if (!Main.PlayerStates.TryGetValue(seer.PlayerId, out PlayerState state) || state.Role is not Penguin pg || pg.AbductVictim == null) return string.Empty;
 
             return $"\u21b9 {(int)(pg.AbductTimer + 1f)}s";
         }
